requirements.txt
pytest
selenium
allure-pytest
pytest-random-order
